% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheap_bootstrap.R
\name{get_all_bootstrap_confidence_intervals}
\alias{get_all_bootstrap_confidence_intervals}
\alias{get_all_bootstrap_confidence_intervals.cheap_bootstrap}
\title{Get all bootstrap confidence intervals}
\usage{
get_all_bootstrap_confidence_intervals(x, ...)

\method{get_all_bootstrap_confidence_intervals}{cheap_bootstrap}(x, ...)
}
\arguments{
\item{x}{An object of class "cheap_bootstrap"}

\item{...}{Not applicable.
Plots the point estimates and confidence intervals as a function of the number of bootstrap samples.}
}
\description{
Get all bootstrap confidence intervals
}
\examples{
utils::data(anorexia, package = "MASS")
## example with a function call
set.seed(123)
x <- function(d) coef(lm(Postwt ~ Prewt + Treat + offset(Prewt), data = d))
cs <- cheap_bootstrap(x, b = 10, data = anorexia)
get_all_bootstrap_confidence_intervals(cs)
}
