% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheap_bootstrap.R
\name{plot.cheap_bootstrap}
\alias{plot.cheap_bootstrap}
\title{Plot method for cheap_bootstrap objects}
\usage{
\method{plot}{cheap_bootstrap}(x, alternate_confidence_interval = NULL, ...)
}
\arguments{
\item{x}{An object of class "cheap_bootstrap"}

\item{alternate_confidence_interval}{Numeric.
Additional confidence interval to be added to the plot.
Must be a data.frame consisting of the columns specified by parameter_column_names
and two additional columns of type numeric.
for each parameter.}

\item{...}{Not applicable.
Plots the point estimates and confidence intervals
as a function of the number of bootstrap samples.}
}
\description{
Plot method for cheap_bootstrap objects
}
