% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cheap_subsampling.R
\name{print.cheap_bootstrap}
\alias{print.cheap_bootstrap}
\title{Print method for cheap_bootstrap objects}
\usage{
\method{print}{cheap_bootstrap}(x, ...)
}
\arguments{
\item{x}{An object of class "cheap_bootstrap"}

\item{...}{Not applicable.
Prints the point estimates and confidence intervals.}
}
\description{
Print method for cheap_bootstrap objects
}
