% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheap_subsampling.R
\name{print.cheap_subsampling}
\alias{print.cheap_subsampling}
\title{Print method for cheap_subsampling objects}
\usage{
\method{print}{cheap_subsampling}(x, ...)
}
\arguments{
\item{x}{An object of class "cheap_subsampling"}

\item{...}{Not applicable.
Prints the point estimates and confidence intervals.}
}
\description{
Print method for cheap_subsampling objects
}
