% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_cheap_bootstrap_plot}
\alias{shiny_cheap_bootstrap_plot}
\title{Shiny app for cheap subsampling}
\usage{
shiny_cheap_bootstrap_plot(x, data = NULL)
}
\arguments{
\item{x}{A function that returns a vector of coefficients
or a model object which saves the call. In the second case, a
coef needs to be defined for 'class(x)'.}

\item{data}{Data set to be used for the computation, if applicable.}
}
\description{
Shiny app for cheap subsampling plot
}
\examples{
\dontrun{
utils::data(anorexia, package = "MASS")
## example with a function call
set.seed(123)
x <- function(d) coef(lm(Postwt ~ Prewt + Treat + offset(Prewt), data = d))
shiny_cheap_bootstrap_plot(x, data = anorexia)
}

}
