% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cheap_subsampling.R
\name{summary.cheap_subsampling}
\alias{summary.cheap_subsampling}
\title{Summary method for cheap_subsampling objects}
\usage{
\method{summary}{cheap_subsampling}(x, ...)
}
\arguments{
\item{x}{An object of class "cheap_subsampling"}

\item{...}{Not applicable.}
}
\value{
Summary table with the point estimates and confidence intervals.
}
\description{
Summary method for cheap_subsampling objects
}
