% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cheap_bootstrap}
\alias{plot.cheap_bootstrap}
\title{Plot method for cheap_bootstrap objects}
\usage{
\method{plot}{cheap_bootstrap}(x, ...)
}
\arguments{
\item{x}{An object of class "cheap_bootstrap"}

\item{...}{Not applicable.
Plots the point estimates and confidence intervals as a function of the number of bootstrap samples.}
}
\description{
Plot method for cheap_bootstrap objects
}
\examples{
\dontrun{
utils::data(anorexia, package = "MASS")
## example with a function call
set.seed(123)
x <- function(d) coef(lm(Postwt ~ Prewt + Treat + offset(Prewt), data = d))
cs <- cheap_bootstrap(x, b = 1000, data = anorexia)
plot(cs)
}

}
