% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cheap_bootstrap.R
\name{summary.cheap_bootstrap}
\alias{summary.cheap_bootstrap}
\title{Summary method for cheap_subsampling objects}
\usage{
\method{summary}{cheap_bootstrap}(object, print = TRUE, ...)
}
\arguments{
\item{object}{An object of class "cheap_subsampling"}

\item{...}{Not applicable.}
}
\value{
Summary table with the point estimates and confidence intervals.
}
\description{
Summary method for cheap_subsampling objects
}
